Half-way to anamorphisms
I've added a notion of anamorphism for each codata type. Ideally, we would have
a Fix datatype to represent all final carriers, in order to implement `ana` once
and for all, but I'm not sure about its final usability (Do we have to deal with
Fx in the final data?) Therefore, I've deployed specific anamorphisms which
exploit the fact that datatypes are functors (instead of manually implementing
unfold) In the near future, we could find ourselves working with datatype
generics. However, this is good enough for our standing priorities.