Added CCS datatype
I've been working long time in this commit, because I wasn't very sure about its
contents. The main target was to embed CCS combinators into CCS coalgebras, but
it turned to be quite more complex than expected. As a result there are many
comments and function versions which I will be deleting in future commits.

It could be interesting to take a look at `CCSRaw` where CCS semantics are
nicely embedded in a *CCS-Coalgebra*. Next tasks will be oriented to migrate
those combinators to `CCSCore`, when possible. `action6` seems to be the first
pillar to take as a reference.